+++ src/commands/format.rs
@@ -1,72 +1,64 @@
 //! Implementation of the `format` subcommand.
 
-use std::ffi::OsStr;
 use std::fs;
 use std::io::IsTerminal;
-use std::io::Read;
-use std::path::Path;
 
 use anyhow::Context;
 use anyhow::Result;
 use anyhow::anyhow;
 use anyhow::bail;
 use clap::Parser;
-use serde::Deserialize;
-use serde::Serialize;
-use walkdir::WalkDir;
-use wdl::ast::Document;
+use clap::Subcommand;
+use tracing::info;
+use tracing::warn;
+use wdl::analysis::Document;
+use wdl::ast::AstNode;
 use wdl::ast::Node;
+use wdl::cli::Analysis;
 use wdl::cli::analysis::Source;
-use wdl::format::Config;
 use wdl::format::Formatter;
 use wdl::format::config::Builder;
 use wdl::format::config::Indent;
 use wdl::format::config::MaxLineLength;
 use wdl::format::element::node::AstNodeFormatExt;
 
+use crate::IGNORE_FILENAME;
 use crate::Mode;
 use crate::emit_diagnostics;
 
 /// Arguments for the `format` subcommand.
 #[derive(Parser, Debug)]
-#[command(
-    author,
-    version,
-    about,
-    after_help = "Use the `--overwrite` option to replace a WDL document or a directory \
-                  containing WDL documents with the formatted source.\nUse the `--check` option \
-                  to verify that a document or a directory containing WDL documents is already \
-                  formatted and print the diff if not."
-)]
+#[command(author, version, about)]
 pub struct Args {
-    /// The path to the local WDL document or directory containing WDL documents
-    /// to format or check.
-    #[arg(value_name = "PATH or DIR")]
-    pub source: Option<Source>,
-
     /// Disables color output.
-    #[arg(long)]
+    #[arg(long, global = true)]
     pub no_color: bool,
 
-    /// The report mode.
-    #[arg(short = 'm', long, value_name = "MODE")]
+    /// The report mode for any emitted diagnostics.
+    #[arg(short = 'm', long, value_name = "MODE", global = true)]
     pub report_mode: Option<Mode>,
 
     /// Use tabs for indentation (default is spaces).
-    #[arg(long)]
+    #[arg(short = 't', long, global = true)]
     pub with_tabs: bool,
 
     /// The number of spaces to use for indentation levels (default is 4).
-    #[arg(long, value_name = "SIZE", conflicts_with = "with_tabs")]
+    #[arg(
+        short,
+        long,
+        value_name = "SIZE",
+        conflicts_with = "with_tabs",
+        global = true
+    )]
     pub indentation_size: Option<usize>,
 
     /// The maximum line length (default is 90).
-    #[arg(long, value_name = "LENGTH")]
+    #[arg(long, value_name = "LENGTH", global = true)]
     pub max_line_length: Option<usize>,
 
-    /// Argument group defining the mode of behavior.
-    #[command(flatten)]
-    pub mode: ModeGroup,
+    /// Subcommand for the `format` command.
+    #[command(subcommand)]
+    pub command: FormatSubcommand,
 }
 
 impl Args {
@@ -87,121 +79,69 @@ impl Args {
     }
 }
 
-/// Argument group defining the mode of behavior
-#[derive(Parser, Debug, Deserialize, Serialize)]
-#[group(required = true, multiple = false)]
-pub struct ModeGroup {
-    /// Overwrite the WDL documents with the formatted versions.
-    #[arg(long, conflicts_with = "check")]
-    pub overwrite: bool,
-
-    /// Check if files are formatted correctly and print diff if not.
-    #[arg(long)]
-    pub check: bool,
+/// Vec of Source arguments (may be empty).
+#[derive(Parser, Debug, Clone)]
+pub struct OptionalSources {
+    /// Sources to format.
+    sources: Vec<Source>,
 }
 
-/// Reads source from the given path.
-///
-/// If the path is simply `-`, the source is read from STDIN.
-fn read_source(path: &Path) -> Result<String> {
-    if path.as_os_str() == "-" {
-        let mut source = String::new();
-        std::io::stdin()
-            .read_to_string(&mut source)
-            .context("failed to read source from STDIN")?;
-        Ok(source)
-    } else {
-        Ok(fs::read_to_string(path).with_context(|| {
-            format!("failed to read source file `{path}`", path = path.display())
-        })?)
-    }
+/// Source argument that is required.
+#[derive(Parser, Debug, Clone)]
+pub struct RequiredSource {
+    /// Source to format.
+    source: Source,
+}
+
+/// Subcommands for the `format` command.
+#[derive(Subcommand, Debug, Clone)]
+pub enum FormatSubcommand {
+    /// Check if files are formatted correctly and print diff if not.
+    Check(OptionalSources),
+
+    /// Format a document and send the result to STDOUT.
+    View(RequiredSource),
+
+    /// Reformat all WDL documents via overwriting.
+    Overwrite(OptionalSources),
 }
 
 /// Formats a document.
-///
-/// If `check_only` is true, checks if the document is formatted correctly and
-/// prints the diff if not then exits. Else will format and overwrite the
-/// document.
-///
-/// If the document failed to parse, this emits the diagnostics and returns
-/// `Ok(count)` of the diagnostics to the caller.
-///
-/// A return value of `Ok(0)` indicates the document was formatted.
 fn format_document(
-    config: Config,
-    path: &Path,
-    report_mode: Mode,
+    formatter: &Formatter,
+    document: &Document,
+    mode: Mode,
     no_color: bool,
-    check_only: bool,
-) -> Result<usize> {
-    let source = read_source(path)?;
-    let (document, diagnostics) = Document::parse(&source);
+) -> Result<(String, String)> {
+    let source = document.root().text().to_string();
+    let diagnostics = document
+        .diagnostics()
+        .iter()
+        .filter(|d| d.severity().is_error())
+        .collect::<Vec<_>>();
     if !diagnostics.is_empty() {
-        emit_diagnostics(
-            path.as_os_str().to_str().expect("path is not UTF-8"),
-            source,
-            &diagnostics,
-            &[],
-            report_mode,
-            no_color,
-        )
-        .context("failed to emit diagnostics")?;
-
-        return Ok(diagnostics.len());
+        let path = document.path();
+        emit_diagnostics(&path, source.clone(), diagnostics, &[], mode, no_color)?;
+        return Err(anyhow!("cannot format a malformed document"));
     }
 
-    let document = Node::Ast(
-        document
-            .ast()
-            .into_v1()
-            .ok_or_else(|| anyhow!("only WDL 1.x documents are currently supported"))?,
-    )
-    .into_format_element();
-
-    let formatter = Formatter::new(config);
-    let formatted = formatter.format(&document)?;
-
-    if check_only {
-        if formatted != source {
-            if !no_color && std::io::stderr().is_terminal() {
-                eprint!(
-                    "{}",
-                    pretty_assertions::StrComparison::new(&source, &formatted)
-                );
-            } else {
-                let diff = similar::TextDiff::from_lines(&source, &formatted);
-                eprint!("{}", diff.unified_diff());
-            }
-            return Ok(1);
-        }
-        println!("`{path}` is formatted correctly", path = path.display());
-        return Ok(0);
-    }
-
-    // Write file because check is not true
-    fs::write(path, formatted)
-        .with_context(|| format!("failed to write `{path}`", path = path.display()))?;
-
-    Ok(0)
+    let ast = document
+        .root()
+        .ast()
+        .into_v1()
+        .expect("only WDL v1.x documents are supported");
+    let element = Node::Ast(ast).into_format_element();
+    Ok((source, formatter.format(&element)?))
 }
 
 /// Runs the `format` command.
-pub fn format(args: Args) -> Result<()> {
-    if let Some(Source::Remote(_)) = args.source {
-        bail!("remote sources are not supported for the `format` command");
-    }
-
-    let source = args.source.unwrap_or_default();
-    let indent = match Indent::try_new(args.with_tabs, args.indentation_size) {
-        Ok(indent) => indent,
-        Err(e) => bail!("failed to create indentation configuration: {}", e),
-    };
+pub async fn format(args: Args) -> Result<()> {
+    let indent = Indent::try_new(args.with_tabs, args.indentation_size)
+        .context("failed to create indentation configuration")?;
 
     let max_line_length = match args.max_line_length {
-        Some(length) => match MaxLineLength::try_new(length) {
-            Ok(max_line_length) => max_line_length,
-            Err(e) => bail!("failed to create max line length configuration: {}", e),
-        },
+        Some(length) => MaxLineLength::try_new(length)
+            .context("failed to create max line length configuration")?,
         None => MaxLineLength::default(),
     };
 
@@ -209,42 +149,178 @@ pub fn format(args: Args) -> Result<()> {
         .indent(indent)
         .max_line_length(max_line_length)
         .build();
+    let formatter = Formatter::new(config);
 
-    let mut diagnostics = 0;
-    if let Source::Directory(path) = source {
-        for entry in WalkDir::new(&path) {
-            let entry = entry.with_context(|| {
-                format!("failed to walk directory `{path}`", path = path.display())
-            })?;
-            let path = entry.path();
-            if !path.is_file() || path.extension().and_then(OsStr::to_str) != Some("wdl") {
-                continue;
+    let mut errors = 0;
+    match args.command {
+        FormatSubcommand::Check(s) => {
+            let mut sources = s.sources;
+            if sources.is_empty() {
+                sources.push(Source::default());
             }
 
-            diagnostics += format_document(
-                config,
-                path,
+            let results = match Analysis::default()
+                .extend_sources(sources.clone())
+                .ignore_filename(Some(IGNORE_FILENAME.to_string()))
+                .run()
+                .await
+            {
+                Ok(results) => results,
+                Err(errors) => {
+                    // SAFETY: this is a non-empty, so it must always have a first
+                    // element.
+                    bail!(errors.into_iter().next().unwrap())
+                }
+            };
+            let sources = sources.iter().collect::<Vec<_>>();
+            let results = results.filter(sources.as_slice()).collect::<Vec<_>>();
+            for result in results {
+                info!("checking `{}`", result.document().path());
+
+                if let Some(err) = result.error() {
+                    errors += 1;
+                    warn!("error analyzing `{}`: {}", result.document().path(), err);
+                    continue;
+                }
+
+                let (source, formatted) = match format_document(
+                    &formatter,
+                    result.document(),
+                    args.report_mode.unwrap_or_default(),
+                    args.no_color,
+                ) {
+                    Ok(r) => r,
+                    Err(e) => {
+                        errors += 1;
+                        warn!(
+                            "skipping format check for `{}`: {e}",
+                            result.document().path()
+                        );
+                        continue;
+                    }
+                };
+                if formatted != source {
+                    warn!("difference in `{}`", result.document().path());
+                    if !args.no_color && std::io::stderr().is_terminal() {
+                        eprint!(
+                            "{}",
+                            pretty_assertions::StrComparison::new(&source, &formatted)
+                        );
+                    } else {
+                        let diff = similar::TextDiff::from_lines(&source, &formatted);
+                        eprint!("{}", diff.unified_diff().header("input", "formatted"));
+                    }
+                    errors += 1;
+                } else {
+                    println!("`{}` is formatted correctly", result.document().path())
+                }
+            }
+        }
+        FormatSubcommand::View(s) => {
+            let source = s.source;
+            match &source {
+                Source::File(_) | Source::Remote(_) => {}
+                Source::Directory(p) => {
+                    bail!(
+                        "the `format view` command does not support formatting directory `{path}`",
+                        path = p.display()
+                    );
+                }
+            };
+
+            let results = match Analysis::default().add_source(source.clone()).run().await {
+                Ok(results) => results,
+                Err(errors) => {
+                    // SAFETY: this is a non-empty, so it must always have a first
+                    // element.
+                    bail!(errors.into_iter().next().unwrap())
+                }
+            };
+            let result = results.filter(&[&source]).next().unwrap();
+
+            if let Some(err) = result.error() {
+                bail!(
+                    "error analyzing `{path}`: {err:#}",
+                    path = result.document().path()
+                );
+            }
+
+            let (_source, formatted) = format_document(
+                &formatter,
+                result.document(),
                 args.report_mode.unwrap_or_default(),
                 args.no_color,
-                args.mode.check,
-            )?;
+            )
+            .with_context(|| {
+                format!(
+                    "could not view document `{path}`",
+                    path = result.document().path()
+                )
+            })?;
+            print!("{}", formatted);
         }
-    } else if let Source::File(path) = source {
-        diagnostics += format_document(
-            config,
-            &path.to_file_path().expect("should be local file path"),
-            args.report_mode.unwrap_or_default(),
-            args.no_color,
-            args.mode.check,
-        )?;
-    } else {
-        unreachable!()
+        FormatSubcommand::Overwrite(s) => {
+            let mut sources = s.sources;
+            if sources.is_empty() {
+                sources.push(Source::default());
+            }
+
+            let results = match Analysis::default()
+                .extend_sources(sources.clone())
+                .ignore_filename(Some(IGNORE_FILENAME.to_string()))
+                .run()
+                .await
+            {
+                Ok(results) => results,
+                Err(errors) => {
+                    // SAFETY: this is a non-empty, so it must always have a first
+                    // element.
+                    bail!(errors.into_iter().next().unwrap())
+                }
+            };
+            let sources = sources.iter().collect::<Vec<_>>();
+            let results = results.filter(sources.as_slice()).collect::<Vec<_>>();
+            for result in results {
+                info!("formatting `{}`", result.document().path());
+
+                if let Some(err) = result.error() {
+                    errors += 1;
+                    warn!(
+                        "error analyzing `{path}`: {err:#}",
+                        path = result.document().path()
+                    );
+                    continue;
+                }
+
+                let (_source, formatted) = match format_document(
+                    &formatter,
+                    result.document(),
+                    args.report_mode.unwrap_or_default(),
+                    args.no_color,
+                ) {
+                    Ok(r) => r,
+                    Err(e) => {
+                        errors += 1;
+                        warn!(
+                            "not overwriting document `{path}` due to error: {e:#}",
+                            path = result.document().path()
+                        );
+                        continue;
+                    }
+                };
+
+                fs::write(result.document().uri().to_file_path().unwrap(), formatted)
+                    .with_context(|| {
+                        format!("failed to overwrite `{}`", result.document().path())
+                    })?;
+            }
+        }
     }
 
-    if diagnostics > 0 {
+    if errors > 0 {
         bail!(
-            "aborting due to previous {diagnostics} diagnostic{s}",
-            s = if diagnostics == 1 { "" } else { "s" }
+            "failing due to previous {errors} error{s}",
+            s = if errors == 1 { "" } else { "s" }
         );
     }